+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Change layers icons to make them a bit more distinct
 - Remove `@mdi` packages in favor of `react-icons`
 - Add ability to control the projection of the map - either globe or mercator
+- Added global state modal to allow editing the global state
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes